+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/DocumentAnalysisClient.java
@@ -184,7 +184,6 @@ public final class DocumentAnalysisClient {
                 finalAnalyzeDocumentOptions.getPages(),
                 finalAnalyzeDocumentOptions.getLocale(),
                 toInnerDocAnalysisFeatures(finalAnalyzeDocumentOptions.getDocumentAnalysisFeatures()),
-                finalAnalyzeDocumentOptions.getQueryFields(),
                 null,
                 documentUrl,
                 finalContext).apply(cxt)),
@@ -298,7 +297,6 @@ public final class DocumentAnalysisClient {
                 finalAnalyzeDocumentOptions.getPages(),
                 finalAnalyzeDocumentOptions.getLocale(),
                 toInnerDocAnalysisFeatures(finalAnalyzeDocumentOptions.getDocumentAnalysisFeatures()),
-                finalAnalyzeDocumentOptions.getQueryFields(),
                 document,
                 null,
                 finalContext).apply(cxt)),
@@ -422,14 +420,13 @@ public final class DocumentAnalysisClient {
     }
 
     private Function<PollingContext<OperationResult>, OperationResult> analyzeActivationOperation(
-        String modelId, List<String> pages, String locale, List<com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentAnalysisFeature> features, List<String> queryFields,
+        String modelId, List<String> pages, String locale, List<com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentAnalysisFeature> features,
         BinaryData document, String documentUrl, Context context) {
         return (pollingContext) ->
             Transforms.toDocumentOperationResult(analyzeDocument(modelId,
                 CoreUtils.isNullOrEmpty(pages) ? null : String.join(",", pages),
                 locale,
                 features,
-                queryFields,
                 document,
                 documentUrl,
                 context)
@@ -460,7 +457,7 @@ public final class DocumentAnalysisClient {
     }
 
     private ResponseBase<DocumentModelsAnalyzeDocumentHeaders, Void> analyzeDocument(String modelId, String pages, String locale,
-                                                                                     List<com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentAnalysisFeature> features, List<String> queryFields,
+                                                                                     List<com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentAnalysisFeature> features,
                                                                                      BinaryData document, String documentUrl, Context context) {
         try {
             if (documentUrl == null) {
@@ -470,7 +467,6 @@ public final class DocumentAnalysisClient {
                     locale,
                     StringIndexType.UTF16CODE_UNIT,
                     features,
-                    queryFields,
                     document,
                     document.getLength(),
                     context);
@@ -480,7 +476,6 @@ public final class DocumentAnalysisClient {
                     locale,
                     StringIndexType.UTF16CODE_UNIT,
                     features,
-                    queryFields,
                     new AnalyzeDocumentRequest().setUrlSource(documentUrl),
                     context);
             }